dev-requirements.txt
-r test-requirements.txt
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
pytest-cov===2.5.1
codecov==2.0.9
flake8==3.4.1
isort==4.2.15
matplotlib==2.1.1
tox==2.9.1
